LPS-171343 Move Util class to api
